pyproject.toml
[build-system]
requires = ["setuptools >= 65.5.0", "setuptools_scm >= 8.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pmod"
authors = [
    {name = "Andy Pangaribuan", email = "iam.pangaribuan@gmail.com"}
]
description = "python module"
version = "0.0.8"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv == 1.0.1",
    "rich == 13.9.3",
    "requests == 2.32.3",
    "grpcio == 1.67.0",
    "grpcio-tools == 1.67.0",
    "psycopg2 == 2.9.10",
    "psycopg2-binary == 2.9.10",
    "tabulate == 0.9.0",
    "packaging == 24.1",
]

[tool.ruff]
exclude = ["__init__.py"]
